backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
google-auth
google-auth-oauthlib
google-auth-httplib2
pydantic[email]
python-dotenv